OpenAI-Azure-Service/requirements.txt
openai==1.55.3
tiktoken==0.8.0
fastapi==0.115.5
uvicorn[standard]==0.32.0
config-client==1.4.0
overrides==7.7.0

# gemini
google-generativeai==0.8.3

langchain-core==0.3.21
langchain==0.3.4
langchain-community==0.3.3
langchain-openai==0.2.3
langchain-aws==0.2.7
langchain-google-genai==2.0.1
langchain-experimental==0.3.2

pypdf==4.3.1
pypdf2==3.0.1
pdfplumber==0.11.4

pymupdf==1.24.13

azure-identity==1.19.0
azure-search-documents==11.5.1
azure-storage-blob==12.24.0
azure-monitor-opentelemetry==1.6.2
azure-keyvault-secrets==4.9.0

python-multipart==0.0.12

pymongo==4.9.0
motor==3.6.0
pydantic-settings==2.6.0
hvac==2.3.0
unstructured[docx,csv,xlsx]

PyJWT==2.9.0

#gabi
pydub
yt-dlp

# para corrigir erro de importação do opentelemetry no configtest para o pytest
setuptools==75.3.0

python-docx==0.8.11
openpyxl==3.1.2

redis==5.2.0
# numpy==1.26.0
